requirements.txt
# OpenTelemetry core packages
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-semantic-conventions>=0.41b0

# Azure Monitor OpenTelemetry exporter
azure-monitor-opentelemetry>=1.0.0